Decouple puzzle model
